requirements.txt
python-dotenv>=0.21.1
beautifulsoup4>=4.12.2
requests>=2.31.0